style(infrastructure): Move `asset-uploader.js` to `test/screenshot/lib/`
The root `test/screenshot/` directory should only contain `run-*.js` files for `npm` commands.

All other JS files should go in `lib/`.

This PR also moves `module.exports` to the top of the file for better readability.
